Remove solidity parser from lib-sourcify, move it back to server.
