+++ packages/lib-sourcify/package.json
@@ -44,7 +44,6 @@
     "@ethersproject/bignumber": "5.8.0",
     "@ethersproject/bytes": "5.8.0",
     "@fairdatasociety/bmt-js": "2.1.0",
-    "@solidity-parser/parser": "0.20.2",
     "bs58": "5.0.0",
     "ethers": "6.15.0",
     "jszip": "3.10.1",
